externalized stack and stackElement
